Add swipe view to dialer app
Summary:
This allows the user to swipe between different dialer pages.

Currently it just uses a `TabBar`:
{F6080958}

But I'd like to have it look kind of like it used to, with text under an icon:
{F6080974}

I don't know how to do this with QML though, I think I you a label and icon in a column layout but I can't find the QML component for just an 'icon'.

Reviewers: mart, #plasma:_mobile

Differential Revision: https://phabricator.kde.org/D14057